BottomTabs 컴포넌트 분리  , ImageButton 이전에 바텀탭에 사용하던 프롭스 삭제
